rewrote intro, first example
Rewrote the beginning of the episode including the Canny filter explanation to
fit skimage better. In particular, added sigma as parameter that should be
determined. Also skimage will convert images to float64 when read with
`as_gray=True`; added a note about this and made sure low and high threshold
values are floats now, too.